requirements.txt

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
itsdangerous==2.1.2
PyJWT==2.8.0
qrcode[pil]==7.4.2
Pillow==10.0.1
reportlab==4.0.5
python-magic==0.4.27
WeasyPrint==60.1
